chore: Move JSON with comments to a new module, b=no-bug, c=scripts, tests
