+++ scripts/run_tests.py
@@ -7,6 +7,7 @@ import sys
 import json
 from pathlib import Path
 from typing import Any
+from json_with_comments import JSONWithCommentsDecoder
 
 IGNORE_PREFS_FILE_IN = os.path.join(
     'src', 'zen', 'tests', 'ignorePrefs.json'
@@ -14,15 +15,6 @@ IGNORE_PREFS_FILE_IN = os.path.join(
 IGNORE_PREFS_FILE_OUT = os.path.join(
     'engine', 'testing', 'mochitest', 'ignorePrefs.json'
 )
-
-
-class JSONWithCommentsDecoder(json.JSONDecoder):
-  def __init__(self, **kw):
-    super().__init__(**kw)
-
-  def decode(self, s: str) -> Any:
-    s = '\n'.join(l for l in s.split('\n') if not l.lstrip(' ').startswith('//'))
-    return super().decode(s)
 
 
 def copy_ignore_prefs():
